refactor(pipeline): peel Types off cycle (12i-1)
Redirect two Types-layer interface imports to their origin leaf files,
removing the two wrong-direction Types->Mediator edges that pulled the
Types domain into the 69-file Mediator import cycle.

- LoginTypes: import IFormAnchor from Mediator/Form/Anchor/AnchorTypes
  (its definition site) instead of the Form/FormAnchor barrel, which is
  itself a cycle member.
- ScrapeDiscoveryTypes: import IDiscoveredEndpoint from
  Mediator/Network/Types/Endpoint (definition site) instead of the
  NetworkDiscoveryTypes barrel.

Both origins are pure type leaves outside the SCC, so the imports
resolve to identical symbols (tsc clean) while the cycle no longer
reaches back into Types. LoginTypes and ScrapeDiscoveryTypes are
peeled off the SCC, shrinking the frozen baseline cycle from 69 to
67 nodes.
